+++ src/opcontrol.cpp
@@ -31,7 +31,7 @@ void opcontrol() {
 		//int dir = hardware::master->get_analog(ANALOG_RIGHT_Y);
 
 		int dir = ((180.0/3.14159265) * atan2(hardware::master.get_analog(ANALOG_LEFT_X), hardware::master.get_analog(ANALOG_LEFT_Y)));
-		int speed = sqrt(pow(hardware::master.get_analog(ANALOG_LEFT_Y),2) + pow(hardware::master.get_analog(ANALOG_LEFT_X),2));
+		float speed = sqrt(pow(hardware::master.get_analog(ANALOG_LEFT_Y),2) + pow(hardware::master.get_analog(ANALOG_LEFT_X),2));
 		hardware::swerve.set_angle(dir);
 		//hardware::drive_mtr.move_velocity(speed);
